Splitting up query config files and tests (#5591)
