+++ tests/fixtures/postgres_fixtures.py
@@ -2,6 +2,7 @@ from typing import Dict, Generator, List
 from uuid import uuid4
 
 import pytest
+from fideslang.models import Dataset as FideslangDataset
 from sqlalchemy.orm import Session
 from sqlalchemy.orm.exc import ObjectDeletedError
 from sqlalchemy_utils.functions import drop_database
@@ -23,6 +24,7 @@ from fides.api.models.sql_models import Dataset as CtlDataset
 from fides.api.models.sql_models import System
 from fides.api.service.connectors import PostgreSQLConnector
 from fides.config import CONFIG
+from tests.ops.test_helpers.dataset_utils import remove_primary_keys
 from tests.ops.test_helpers.db_utils import seed_postgres_data
 
 from .application_fixtures import integration_secrets
@@ -97,6 +99,34 @@ def postgres_example_test_dataset_config_read_access(
     fides_key = postgres_dataset["fides_key"]
 
     ctl_dataset = CtlDataset.create_from_dataset_dict(db, postgres_dataset)
+
+    dataset = DatasetConfig.create(
+        db=db,
+        data={
+            "connection_config_id": read_connection_config.id,
+            "fides_key": fides_key,
+            "ctl_dataset_id": ctl_dataset.id,
+        },
+    )
+    yield dataset
+    dataset.delete(db=db)
+    ctl_dataset.delete(db=db)
+
+
+@pytest.fixture
+def postgres_example_test_dataset_config_read_access_without_primary_keys(
+    read_connection_config: ConnectionConfig,
+    db: Session,
+    example_datasets: List[Dict],
+) -> Generator:
+    postgres_dataset = example_datasets[0]
+    fides_key = postgres_dataset["fides_key"]
+
+    dataset = FideslangDataset(**postgres_dataset)
+    updated_dataset = remove_primary_keys(dataset)
+    ctl_dataset = CtlDataset.create_from_dataset_dict(
+        db, updated_dataset.model_dump(mode="json")
+    )
 
     dataset = DatasetConfig.create(
         db=db,
