+++ tests/ops/service/connectors/test_snowflake_query_config.py
@@ -150,7 +150,7 @@ class TestSnowflakeQueryConfig:
         )
         assert (
             str(update_stmt)
-            == 'UPDATE "address" SET "city" = :city, "house" = :house, "state" = :state, "street" = :street, "zip" = :zip WHERE "id" = :id'
+            == 'UPDATE "address" SET "city" = :masked_city, "house" = :masked_house, "state" = :masked_state, "street" = :masked_street, "zip" = :masked_zip WHERE "id" = :id'
         )
 
     def test_generate_namespaced_update_stmt(
@@ -191,5 +191,5 @@ class TestSnowflakeQueryConfig:
         )
         assert (
             str(update_stmt)
-            == 'UPDATE "FIDESOPS_TEST"."TEST"."address" SET "city" = :city, "house" = :house, "state" = :state, "street" = :street, "zip" = :zip WHERE "id" = :id'
+            == 'UPDATE "FIDESOPS_TEST"."TEST"."address" SET "city" = :masked_city, "house" = :masked_house, "state" = :masked_state, "street" = :masked_street, "zip" = :masked_zip WHERE "id" = :id'
         )
