+++ tests/ops/service/dataset/example_datasets/multiple_identities.yml
@@ -16,8 +16,6 @@ dataset:
               data_type: string
           - name: id
             data_categories: [user.unique_id]
-            fides_meta:
-              primary_key: True
           - name: name
             data_categories: [user.name]
             fides_meta:
