+++ tests/ops/service/dataset/example_datasets/multiple_identities_with_external_dependency.yml
@@ -32,7 +32,5 @@ dataset:
                   direction: from
           - name: id
             data_categories: [system.operations]
-            fides_meta:
-              primary_key: True
           - name: shipping_address_id
             data_categories: [system.operations]
