+++ tests/ops/service/dataset/example_datasets/no_identities.yml
@@ -13,8 +13,6 @@ dataset:
             data_categories: [user.contact.email]
           - name: id
             data_categories: [user.unique_id]
-            fides_meta:
-              primary_key: True
           - name: name
             data_categories: [user.name]
             fides_meta:
